// src/engine/types.ts
export type HiveEngineMethodName = 'getLatestBlockInfo' | 'getContract' | 'findOne' | 'find' | 'getLatestBlockInfo' | 'getBlockInfo' | 'getTransactionInfo' | 'getStatus';

export type HiveEngineAssetSymbol = 'SWAP.HIVE' | 'BEE' | 'DEC' | 'ALPHA' | 'BETA' | 'UNTAMED' | 'CHAOS' | string;

export interface HiveEngineRequest {
    method: HiveEngineMethodName;
    params?: Record<string, any>;
}